Fix jest circular references
